// frontend/src/lib/api/products.ts
export type ProductCategory =
  | "колбаса вареная"
  | "колбаса полукопченая"
  | "колбаса сырокопченая"
  | "колбаса сыровяленая"
  | "мясо"
  | "птица"
  | "рыба горячего копчения"
  | "рыба холодного копчения"
  | "рыба электростатического копчения"
  | "сыр"
  | "сало"
  | "масло"
  | "снеки"
  | "прочее";